+++ actions/config/ios/devices.js
@@ -73,14 +73,19 @@ function printDevices(title, msg) {
 }
 
 function list(config, verbose) {
-    if(config)
-        return listDeviceInProvisioningWithInfo(config, verbose)
-            .then(function (provision) {
+    var cwd = process.cwd(),
+        p = config ? listDeviceInProvisioningWithInfo(config, verbose).then(function (provision) {
                 var title = format("Provisioning Profile %s with Type: %s", provision.name, provision.type),
                     msg = format("\nDevices in configuration: %s", config);
                 printDevices(title, msg)(provision.devices);
-            });
-    else return listDevice(verbose).then(printDevices("\nAll Devices :"));
+            }) : listDevice(verbose).then(printDevices("\nAll Devices :"));
+    return p.then(function (val) {
+        process.chdir(cwd);
+        return val;
+    }, function (err) {
+        process.chdir(cwd);
+        throw err;
+    });
 }
 
 function add(name, uuid, verbose) {
@@ -97,70 +102,84 @@ function add(name, uuid, verbose) {
 }
 
 function attach(uuid, config, verbose) {
-    return tarifaFile.parse(pathHelper.root())
-        .then(function(localSettings) {
-            var id = localSettings.deploy.apple_id,
-                team = localSettings.deploy.apple_developer_team,
-                conf = localSettings.configurations.ios[config];
+    var cwd = process.cwd();
+    process.chdir(pathHelper.root());
+    return tarifaFile.parse(pathHelper.root()).then(function(localSettings) {
+        var id = localSettings.deploy.apple_id,
+            team = localSettings.deploy.apple_developer_team,
+            conf = localSettings.configurations.ios[config];
 
-            return askPassword().then(function (password) {
-                return getDevices(id, team, password, verbose).then(function (devices) {
-                    var rslt = devices.filter(function (id) { return id === uuid; }),
-                        profile_path = conf.provisioning_profile_path,
-                        profile_name = conf.provisioning_profile_name;
+        return askPassword().then(function (password) {
+            return getDevices(id, team, password, verbose).then(function (devices) {
+                var rslt = devices.filter(function (id) { return id === uuid; }),
+                    profile_path = conf.provisioning_profile_path,
+                    profile_name = conf.provisioning_profile_name;
 
-                    if(rslt.length) {
-                        if(verbose) print.success('device already in developer center');
+                if(rslt.length) {
+                    if(verbose) print.success('device already in developer center');
+                    return provisioningManager.add(id, team, password, uuid, profile_path, devices, verbose).then(function () {
+                        return downloadProvisioning(id, team, password, profile_name, profile_path, verbose);
+                    });
+                }
+                else {
+                    if(verbose) print.success('device not in developer center');
+                    return askDeviceName().then(function (name) {
+                        return addDevice(id, team, password, name, uuid, verbose).then(function (output) {
+                            devices.push({ name:name, uuid:uuid, enabled:true });
+                            if(verbose) print(output);
+                        });
+                    }).then(function () {
                         return provisioningManager.add(id, team, password, uuid, profile_path, devices, verbose).then(function () {
                             return downloadProvisioning(id, team, password, profile_name, profile_path, verbose);
                         });
-                    }
-                    else {
-                        if(verbose) print.success('device not in developer center');
-                        return askDeviceName().then(function (name) {
-                            return addDevice(id, team, password, name, uuid, verbose).then(function (output) {
-                                devices.push({ name:name, uuid:uuid, enabled:true });
-                                if(verbose) print(output);
-                            });
-                        }).then(function () {
-                            return provisioningManager.add(id, team, password, uuid, profile_path, devices, verbose).then(function () {
-                                return downloadProvisioning(id, team, password, profile_name, profile_path, verbose);
-                            });
-                        });
-                    }
-                });
+                    });
+                }
             });
         });
+    }).then(function (val) {
+        process.chdir(cwd);
+        return val;
+    }, function (err) {
+        process.chdir(cwd);
+        throw err;
+    });
 }
 
 function detach(uuid, config, verbose) {
-    return tarifaFile.parse(pathHelper.root())
-        .then(function (localSettings) {
-            var conf = localSettings.configurations.ios[config];
-            if(!conf)
-                return Q.reject('configuration not found');
-            if(!conf.provisioning_profile_path)
-                return Q.reject('no provisioning_profile_path attribute in configuration');
-            if(!conf.provisioning_profile_name)
-                return Q.reject('no provisioning_profile_name attribute in configuration');
+    var cwd = process.cwd();
+    process.chdir(pathHelper.root());
+    return tarifaFile.parse(pathHelper.root()).then(function (localSettings) {
+        var conf = localSettings.configurations.ios[config];
+        if(!conf)
+            return Q.reject('configuration not found');
+        if(!conf.provisioning_profile_path)
+            return Q.reject('no provisioning_profile_path attribute in configuration');
+        if(!conf.provisioning_profile_name)
+            return Q.reject('no provisioning_profile_name attribute in configuration');
 
-            return askPassword().then(function (password) {
-                var profile_path = conf.provisioning_profile_path,
-                    profile_name = conf.provisioning_profile_name,
-                    id = localSettings.deploy.apple_id,
-                    team = localSettings.deploy.apple_developer_team;
-                return parseProvisionFile(profile_path)
-                    .then(function (provision) {
-                        if(provision.uuids.indexOf(uuid) < 0)
-                            return Q.reject('device is not included in the provisioning file!');
-                        return getDevices(id, team, password, verbose).then(function (devices) {
-                            return provisioningManager.remove(id, team, password, uuid, profile_path, devices, verbose);
-                        });
-                    }).then(function () {
-                        return downloadProvisioning(id, team, password, profile_name, profile_path, verbose);
+        return askPassword().then(function (password) {
+            var profile_path = conf.provisioning_profile_path,
+                profile_name = conf.provisioning_profile_name,
+                id = localSettings.deploy.apple_id,
+                team = localSettings.deploy.apple_developer_team;
+            return parseProvisionFile(profile_path)
+                .then(function (provision) {
+                    if(provision.uuids.indexOf(uuid) < 0)
+                        return Q.reject('device is not included in the provisioning file!');
+                    return getDevices(id, team, password, verbose).then(function (devices) {
+                        return provisioningManager.remove(id, team, password, uuid, profile_path, devices, verbose);
                     });
-            });
+                }).then(function () {
+                    return downloadProvisioning(id, team, password, profile_name, profile_path, verbose);
+                });
         });
+    }).then(function (val) {
+        process.chdir(cwd);
+        return val;
+    }, function (err) {
+        process.chdir(cwd);
+        throw err;
+    });
 }
 
 module.exports = {