+++ actions/create/tasks/platforms.js
@@ -6,15 +6,8 @@ var Q = require('q'),
     platformsLib = require('../../../lib/cordova/platforms');
 
 module.exports = function (response) {
-    var platforms = response.platforms.filter(function (platform) { return platform != 'web'; }),
-        cwd = process.cwd();
-
-    if (!platforms.length) return response;
-
-    process.chdir(response.path);
-
-    return platformsLib.add(platforms, response.options.verbose).then(function() {
-        process.chdir(cwd);
+    if (!response.platforms.length) return response;
+    return platformsLib.add(response.path, response.platforms, response.options.verbose).then(function() {
         return response;
     });
 };