+++ actions/run/index.js
@@ -20,7 +20,7 @@ var Q = require('q'),
         ],
         ios : [ './tasks/ios/install' ],
         wp8: [ './tasks/wp8/install' ],
-        web: [ './tasks/web/open' ]
+        browser: [ './tasks/browser/open' ]
     };
 
 var runƒ = function (conf) {