+++ CHANGES.md
@@ -1,4 +1,37 @@
-## 0.2.5 10/09/2014¬
+## 0.3.0 10/XX/2014
+
+### upgrade a project from 0.2.5 to 0.3.0
+
+- remove the `web` platform and the according configuration in the tarifa.json
+- execute the command `tarifa update --verbose` to update cordova plugins and platforms
+- execute the command `tarifa platform add browser --verbose` to add the browser platform
+- in the tarifa.json file, rename the `check.web` key to `check.browser` and ensure the
+corresponding value is `'./project/bin/check_browser.js'`
+
+### changes
+
+- update cordova-lib to 4.0.0
+- replace the `web` platform with the new `browser` platform
+- `tarifa info` adding current project cordova-$platform versions
+- adding tests: `npm test` and `npm run all`
+- www project output can be changed in any configuration with
+the `project_output` attribute
+- warn user if name given in tarifa file does not match the cordova project's name
+- icons and splashscreens folder `images` can be overwritten with the `assets_path`
+attribute in the tarifa.json file
+- the cordova attribute in tarifa.json can be overwritten in any configuration
+- new command `tarifa update` for updating default plugins and cordova platforms
+- call tarifa in any project subdirectory
+- adding a wp8 gitignore
+- cleaning gradle build cache on `tarifa clean`
+- update .gitignore for gradle builds
+- ability to overwrite chrome path on linux and windows in user configstore yaml file
+- be able to choose `all` devices on `tarifa run`
+- speed up the cli by not requiring all actions on start
+- support 9patch splashscreens on android
+- `tarifa platform` supports following format for add action: `$platform@version`
+
+## 0.2.5 10/09/2014
 
 - remove postinstall scripts